fix: Fixed Tool having a random color when creating a new project and camera not moving with right click
The problem with Tool was that it has the member mCurrentColor which is a pointer that points to the used color (mColor1 or mColor2). During the assignment, the value of the pointer from another Tool object would just be copied and mCurrentColor would point to an invalid address. The issue with camera was that I didn't set canvas dimension that it used in CursosPosCallback. Note to self: don't use a member pointer that points to other members!